+++ src/main/kotlin/com/zepben/evolve/cim/iec61968/common/StreetAddress.kt
@@ -15,5 +15,10 @@ package com.zepben.evolve.cim.iec61968.common
  * @property poBox Post office box.
  * @property streetDetail Street detail.
  */
-data class StreetAddress(val postalCode: String, val townDetail: TownDetail?, val poBox: String, val streetDetail: StreetDetail?)
+data class StreetAddress(
+    val postalCode: String = "1234",
+    val townDetail: TownDetail? = TownDetail("name", "stateOrProvince"),
+    val poBox: String = "5678",
+    val streetDetail: StreetDetail? = StreetDetail("buildingName", "floorIdentification", "name", "number", "suiteNumber", "type", "displayAddress")
+)