+++ src/main/kotlin/com/zepben/evolve/database/sqlite/tables/iec61968/common/TableStreetAddresses.kt
@@ -15,12 +15,12 @@ abstract class TableStreetAddresses : TableTownDetails() {
 
     val POSTAL_CODE = Column(++columnIndex, "postal_code", "TEXT", NOT_NULL)
     val PO_BOX = Column(++columnIndex, "po_box", "TEXT", NOT_NULL)
-    val BUILDING_NAME = Column(++columnIndex, "building_name", "TEXT", Column.Nullable.NULL)
-    val FLOOR_IDENTIFICATION = Column(++columnIndex, "floor_identification", "TEXT", Column.Nullable.NULL)
-    val NAME = Column(++columnIndex, "name", "TEXT", Column.Nullable.NULL)
-    val NUMBER = Column(++columnIndex, "number", "TEXT", Column.Nullable.NULL)
-    val SUITE_NUMBER = Column(++columnIndex, "suite_number", "TEXT", Column.Nullable.NULL)
-    val TYPE = Column(++columnIndex, "type", "TEXT", Column.Nullable.NULL)
-    val DISPLAY_ADDRESS = Column(++columnIndex, "display_address", "TEXT", Column.Nullable.NULL)
+    val BUILDING_NAME = Column(++columnIndex, "building_name", "TEXT", NOT_NULL)
+    val FLOOR_IDENTIFICATION = Column(++columnIndex, "floor_identification", "TEXT", NOT_NULL)
+    val NAME = Column(++columnIndex, "name", "TEXT", NOT_NULL)
+    val NUMBER = Column(++columnIndex, "number", "TEXT", NOT_NULL)
+    val SUITE_NUMBER = Column(++columnIndex, "suite_number", "TEXT", NOT_NULL)
+    val TYPE = Column(++columnIndex, "type", "TEXT", NOT_NULL)
+    val DISPLAY_ADDRESS = Column(++columnIndex, "display_address", "TEXT", NOT_NULL)
 
 }