+++ changelog.md
@@ -10,10 +10,12 @@
 * Added the following CIM classes/enums:
   * `TransformerConstructionKind`
   * `TransformerFunctionKind`
+  * `StreetDetail`
 
 ##### Enhancements
 
 * Added fields to `PowerTransformer` to define `constructionKind` and `function`.
+* Added fields to `StreetAddress` to define `poBox` and `streetDetail`.
 
 ##### Fixes